# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "B2B/B2C E-Commerce API Documentation",
        "description": "API documentation for a hybrid B2B/B2C e-commerce platform for occupational health and safety clothing",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b2bb2c.mutfakyapim.net",
            "description": "Production API Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get current user's cart",
                "description": "Retrieve the current cart for authenticated or guest users",
                "operationId": "getCart",
                "responses": {
                    "200": {
                        "description": "Cart retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cart": {
                                                    "type": "object"
                                                },
                                                "summary": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear entire cart",
                "description": "Remove all items from the user's cart",
                "operationId": "clearCart",
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully"
                    },
                    "500": {
                        "description": "Failed to clear cart"
                    }
                }
            }
        },
        "/api/v1/cart/items": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add item to cart",
                "description": "Add a product variant to the user's cart",
                "operationId": "addItemToCart",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_variant_id"
                                ],
                                "properties": {
                                    "product_variant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added to cart successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item added to cart successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to add item to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to add item to cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cart/items/{item}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update item quantity in cart",
                "description": "Update the quantity of an existing cart item",
                "operationId": "updateCartItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Cart item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item updated successfully"
                    },
                    "400": {
                        "description": "Failed to update cart item"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove item from cart",
                "description": "Remove an item from the user's cart",
                "operationId": "removeCartItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Cart item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed from cart successfully"
                    },
                    "400": {
                        "description": "Failed to remove cart item"
                    }
                }
            }
        },
        "/api/v1/cart/summary": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get cart summary",
                "description": "Retrieve cart summary with pricing information",
                "operationId": "getCartSummary",
                "responses": {
                    "200": {
                        "description": "Cart summary retrieved successfully"
                    },
                    "500": {
                        "description": "Failed to retrieve cart summary"
                    }
                }
            }
        },
        "/api/v1/cart/migrate": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Migrate guest cart to authenticated user",
                "description": "Transfer items from guest cart to authenticated user cart",
                "operationId": "migrateGuestCart",
                "responses": {
                    "200": {
                        "description": "Guest cart migrated successfully"
                    },
                    "401": {
                        "description": "User must be authenticated to migrate cart"
                    },
                    "500": {
                        "description": "Failed to migrate guest cart"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    },
...
}